[back] all backlog done or undone api complete and test success

+++ graduation-design-backlog8002/src/main/java/com/jiac/backlog/controller/BacklogController.java
@@ -2,10 +2,7 @@ package com.jiac.backlog.controller;
 
 import com.jiac.backlog.dto.BacklogDto;
 import com.jiac.backlog.feign.UserFeign;
-import com.jiac.backlog.request.AddBacklogRequest;
-import com.jiac.backlog.request.BacklogDeleteRequest;
-import com.jiac.backlog.request.BacklogDoneRequest;
-import com.jiac.backlog.request.BacklogUndoneRequest;
+import com.jiac.backlog.request.*;
 import com.jiac.backlog.service.BacklogService;
 import com.jiac.backlog.vo.BacklogVo;
 import com.jiac.common.entity.Backlog;
@@ -42,7 +39,7 @@ public class BacklogController {
         // 先判断该用户是否存在
         CommonType<Boolean> booleanCommonType = userFeign.userExist(username);
         Boolean userExist = booleanCommonType.getData();
-        if(!userExist) {
+        if(userExist == null) {
             return CommonType.fail(ErrorEnum.USER_NOT_EXIST);
         }
         // 通过service 添加待办事项
@@ -68,6 +65,11 @@ public class BacklogController {
     @ResponseBody
     @GetMapping("/done")
     public CommonType<BacklogVo> done(@RequestParam("id") String id, @RequestParam("username") String username) {
+        // 先判断用户是否存在
+        Boolean userExist = userFeign.userExist(username).getData();
+        if(userExist == null) {
+            return CommonType.fail(ErrorEnum.USER_NOT_EXIST);
+        }
         BacklogDoneRequest request = BacklogDoneRequest.of(id, username);
         BacklogDto done = backlogService.done(request);
         return CommonType.success(BacklogVo.of(done), "修改成功");
@@ -76,6 +78,11 @@ public class BacklogController {
     @ResponseBody
     @GetMapping("/undone")
     public CommonType<BacklogVo> undone(@RequestParam("id") String id, @RequestParam("username") String username) {
+        // 先判断用户是否存在
+        Boolean userExist = userFeign.userExist(username).getData();
+        if(userExist == null) {
+            return CommonType.fail(ErrorEnum.USER_NOT_EXIST);
+        }
         BacklogUndoneRequest request = BacklogUndoneRequest.of(id, username);
         BacklogDto undone = backlogService.undone(request);
         return CommonType.success(BacklogVo.of(undone), "修改成功");
@@ -84,8 +91,26 @@ public class BacklogController {
     @ResponseBody
     @PostMapping("/deleteBacklog")
     public CommonType<BacklogVo> deleteBacklog(@RequestParam("id") String id, @RequestParam("username") String username) {
+        // 先判断用户是否存在
+        Boolean userExist = userFeign.userExist(username).getData();
+        if(userExist == null) {
+            return CommonType.fail(ErrorEnum.USER_NOT_EXIST);
+        }
         BacklogDeleteRequest request = BacklogDeleteRequest.of(id, username);
         BacklogDto backlogDto = backlogService.deleteBacklog(request);
         return CommonType.success(BacklogVo.of(backlogDto), "删除成功");
+    }
+
+    @ResponseBody
+    @GetMapping("/checkAllOrNone")
+    public CommonType<Boolean> checkAllOrNone(@RequestParam("done") Boolean done, @RequestParam("username") String username) {
+        // 先判断用户是否存在
+        Boolean userExist = userFeign.userExist(username).getData();
+        if(userExist == null) {
+            return CommonType.fail(ErrorEnum.USER_NOT_EXIST);
+        }
+        BacklogCheckAllOrNoneRequest request = BacklogCheckAllOrNoneRequest.of(done, username);
+        Boolean success = backlogService.checkAllOrNone(request);
+        return CommonType.success(success, "修改成功");
     }
 }
